# docs/contratos/openapi.yaml
openapi: 3.0.3
info:
  title: MS1 - Pasajeros / Tickets API
  version: 0.1.0
  description: >
    Microservicio de pasajeros y tickets del Aeropuerto Internacional Jorge Chávez.
    Base pública /api/pasajeros (nginx -> puerto 8001). Consume a MS2 (vuelos) via REST.
servers:
  - url: http://localhost:8001
    description: Local (docker-compose)
...
paths:
...
  /pasajeros/{id_persona}/tickets:
    get:
      tags: [Tickets]
      summary: Tickets de un pasajero
      operationId: ticketsDePasajero
      parameters:
        - $ref: '#/components/parameters/IdPersona'
      responses:
        '200':
          description: Tickets del pasajero
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
        '404':
          $ref: '#/components/responses/Error404'
...
  /tickets:
    post:
      tags: [Tickets]
      summary: Emitir ticket (valida vuelo contra MS2)
      operationId: emitirTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearTicketRequest'
      responses:
        '201':
          description: Ticket emitido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
              example:
                id_ticket: 1
                precio: 185.00
                fecha_emision: '2026-09-08'
                estado_boarding: Emitido
                id_vuelo: 12345
                id_persona: 123456
        '400':
          $ref: '#/components/responses/Error400'
        '422':
          description: Vuelo inexistente o cancelado en MS2; o persona inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                error:
                  code: VUELO_NO_EXISTE
                  message: El vuelo 30000 no existe en MS2
        '502':
          $ref: '#/components/responses/Error502'
    get:
      tags: [Tickets]
      summary: Buscar tickets
      operationId: buscarTickets
      parameters:
        - in: query
          name: vuelo_id
          schema:
            type: integer
          description: Filtrar por vuelo (soft MS2)
      responses:
        '200':
          description: Lista de tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
  /tickets/{id_ticket}:
    get:
      tags: [Tickets]
      summary: Obtener ticket por ID
      operationId: obtenerTicket
      parameters:
        - $ref: '#/components/parameters/IdTicket'
      responses:
        '200':
          description: Ticket encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '404':
          $ref: '#/components/responses/Error404'
...
components:
  parameters:
    IdPersona:
      name: id_persona
      in: path
      required: true
      schema:
        type: integer
      description: ID de la persona/pasajero
    IdTicket:
      name: id_ticket
      in: path
      required: true
      schema:
        type: integer
      description: ID del ticket
  responses:
    Error400:
      description: Peticion invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Error404:
      description: No encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    Error502:
      description: Error al consumir MS2
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
...
    EstadoBoarding:
      type: string
      enum: [Emitido, Check-in, Embarcado, No-show, Cancelado]
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: DUPLICADO
            message:
              type: string
              example: Ya existe un pasajero con ese documento
...
    CrearTicketRequest:
      type: object
      required: [precio, fecha_emision, id_vuelo, id_persona]
      properties:
        precio:
          type: number
          format: double
          example: 185.00
        fecha_emision:
          type: string
          format: date
          example: '2026-09-08'
        id_vuelo:
          type: integer
          description: Soft ref a MS2 (1 - 25000)
        id_persona:
          type: integer
          description: FK a persona
    Ticket:
      type: object
      properties:
        id_ticket:
          type: integer
        precio:
          type: number
          format: double
        fecha_emision:
          type: string
          format: date
        estado_boarding:
          $ref: '#/components/schemas/EstadoBoarding'
        id_vuelo:
          type: integer
        id_persona:
          type: integer
...